added double value

+++ src/main/kotlin/com/isotop/storage/dto/response/ContainerResponse.kt
@@ -18,5 +18,5 @@ data class ContainerResponse(
     val openSourceCount: Int?,
     val openSourceRest: Int?,
     val storageCode: Int?,
-    val sourceActivity: Int?
+    val sourceActivity: Double?
 )
